+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [0.2.1] - 2019-02-24
+### Added
+- `lexical-context`: how keys or vals are processed can be tuned with the
+  `:keys` and `:vals` options. Possible values: `:quoted`, `:doubly-quoted`
+  or `:evaled`.
+
 ## [0.2.0] - 2018-11-25
 ### Added
 - `lexical-resolve`.